requirement.txt
pytest
playwright
pytest-playwright
allure-pytest
pytest-html